requirements.txt
dnspython==2.2.1
selenium==3.141.0
urllib3==1.26.19
aiohttp==3.8.6
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0 
charset-normalizer==2.0.12
frozenlist==1.2.0
idna==3.7 
idna-ssl==1.1.0 
maxminddb==2.2.0 
multidict==5.2.0 
requests==2.27.1 
typing-extensions==4.1.1 
yarl==1.7.2 
geoip2==4.6.0
termcolor==1.1.0
PyYAML==5.1.2